+++ src/Akka.Hosting/AkkaHostedService.cs
@@ -1,7 +1,10 @@
-﻿using System.Threading;
+﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using Akka.Actor;
+using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
 namespace Akka.Hosting
 {
@@ -11,36 +14,60 @@ namespace Akka.Hosting
     internal sealed class AkkaHostedService : IHostedService
     {
         private ActorSystem _actorSystem;
+        private CoordinatedShutdown _coordinatedShutdown; // grab a reference to CoordinatedShutdown early
+        private readonly IServiceProvider _serviceProvider;
         private readonly AkkaConfigurationBuilder _configurationBuilder;
         private readonly IHostApplicationLifetime _hostApplicationLifetime;
+        private readonly ILogger<AkkaHostedService> _logger;
 
-        public AkkaHostedService(AkkaConfigurationBuilder configurationBuilder, IHostApplicationLifetime hostApplicationLifetime)
+        public AkkaHostedService(AkkaConfigurationBuilder configurationBuilder, IServiceProvider serviceProvider,
+            ILogger<AkkaHostedService> logger, IHostApplicationLifetime applicationLifetime)
         {
             _configurationBuilder = configurationBuilder;
-            _hostApplicationLifetime = hostApplicationLifetime;
+            _hostApplicationLifetime = applicationLifetime;
+            _serviceProvider = serviceProvider;
+            _logger = logger;
         }
 
         public async Task StartAsync(CancellationToken cancellationToken)
         {
-            _actorSystem = await _configurationBuilder.StartAsync();
-
-            async Task TerminationHook()
+            try
             {
-                await _actorSystem.WhenTerminated.ConfigureAwait(false);
+                _actorSystem = _serviceProvider.GetRequiredService<ActorSystem>();
+                _coordinatedShutdown = CoordinatedShutdown.Get(_actorSystem);
+                await _configurationBuilder.StartAsync(_actorSystem);
+
+                async Task TerminationHook()
+                {
+                    await _actorSystem.WhenTerminated.ConfigureAwait(false);
+                    _hostApplicationLifetime.StopApplication();
+                }
+
+                // terminate the application if the Sys is terminated first
+                // this can happen in instances such as Akka.Cluster membership changes
+#pragma warning disable CS4014
+                TerminationHook();
+#pragma warning restore CS4014
+            }
+            catch (Exception ex)
+            {
+                _logger.Log(LogLevel.Critical, ex, "Unable to start AkkaHostedService - shutting down application");
                 _hostApplicationLifetime.StopApplication();
             }
-
-            // terminate the application if the Sys is terminated first
-            // this can happen in instances such as Akka.Cluster membership changes
-#pragma warning disable CS4014
-            TerminationHook();
-#pragma warning restore CS4014
         }
 
         public async Task StopAsync(CancellationToken cancellationToken)
         {
+            // ActorSystem may have failed to start - skip shutdown sequence if that's the case
+            // so error message doesn't get conflated.
+            if (_coordinatedShutdown == null)
+            {
+                return;
+            }
+            
             // run full CoordinatedShutdown on the Sys
-            await CoordinatedShutdown.Get(_actorSystem).Run(CoordinatedShutdown.ClrExitReason.Instance).ConfigureAwait(false);
+            await _coordinatedShutdown.Run(CoordinatedShutdown.ClrExitReason.Instance)
+                .ConfigureAwait(false);
         }
     }
 }