+++ RELEASE_NOTES.md
@@ -1,2 +1,5 @@
-## [0.1.0] / 09 March 2022
-- Initial release of Akka.Hosting, Akka.Remote.Hosting, Akka.Cluster.Hosting, and Akka.Persistence.SqlServer.Hosting
+## [0.3.0] / 24 May 2022
+- [Add interfaces for the `ActorRegistry` to allow mocking in tests](https://github.com/akkadotnet/Akka.Hosting/pull/42)
+- [Fixed: Akka.Hosting NRE upon shutdown](https://github.com/akkadotnet/Akka.Hosting/pull/49)
+- [added throw-able `ActorRegistry.Register` method](https://github.com/akkadotnet/Akka.Hosting/pull/50)
+- [Akka.Cluster.Hosting: adding `ClusterSingleton` hosting methods](https://github.com/akkadotnet/Akka.Hosting/pull/51)